# docs/swagger.yaml
basePath: /
definitions:
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
...
  dto.CreateDeviceRequest:
    properties:
      organization_id:
        type: integer
      serial_number:
        type: string
    type: object
...
  dto.DeviceResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      organization_id:
        type: integer
      serial_number:
        type: string
      status:
        $ref: '#/definitions/model.DeviceStatus'
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  dto.RegisterOrganizationRequest:
    properties:
      admin_email:
        type: string
      admin_name:
        type: string
      admin_password:
        type: string
      organization_name:
        type: string
    type: object
...
  model.AuthResult:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.DeviceStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - MAINTENANCE
    type: string
    x-enum-varnames:
    - DeviceStatusActive
    - DeviceStatusInactive
    - DeviceStatusMaintenance
...
host: localhost:8080
info:
  contact: {}
  description: Бэкенд системы мониторинга и телематики автопарка
  title: FleetTrack API
  version: "1.0"
paths:
...
  /devices:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeviceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Зарегистрировать GPS-трекер
      tags:
      - devices
  /devices/{id}:
    delete:
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeviceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить трекер
      tags:
      - devices
    get:
      parameters:
      - description: ID устройства
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeviceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить трекер по ID
      tags:
      - devices
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Учётные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.AuthResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Рефреш токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Выход из системы
      tags:
      - auth
...
  /refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Рефреш токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.AuthResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновление токенов доступа и рефреша
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные организации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.AuthResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Регистрация компании
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"